Move address to libmutt and merge it with mutt_address
