+++ Makefile.autosetup
@@ -60,11 +60,11 @@ NEOMUTTOBJS=	mutt_account.o addrbook.o alias.o attach.o bcache.o body.o \
 		curs_main.o edit.o editmsg.o enter.o envelope.o filter.o \
 		flags.o from.o group.o handler.o hdrline.o \
 		header.o help.o history.o hook.o init.o keymap.o main.o \
-		mbox.o menu.o mh.o muttlib.o mutt_address.o \
+		mbox.o menu.o mh.o muttlib.o \
 		mutt_socket.o tags.o mx.o \
 		newsrc.o nntp.o pager.o parse.o pattern.o pop.o \
 		pop_auth.o pop_lib.o postpone.o query.o recvattach.o recvcmd.o \
-		rfc1524.o rfc2047.o rfc2231.o rfc3676.o address.o \
+		rfc1524.o rfc2047.o rfc2231.o rfc3676.o \
 		safe_asprintf.o score.o send.o sendlib.o sidebar.o mutt_signal.o \
 		resize.o smtp.o sort.o state.o status.o system.o thread.o url.o \
 		version.o
@@ -87,8 +87,8 @@ ALLOBJS+=	$(NEOMUTTOBJS)
 ###############################################################################
 # libmutt
 LIBMUTT=	libmutt.a
-LIBMUTTOBJS=	mutt/base64.o mutt/buffer.o mutt/charset.o mutt/date.o \
-		mutt/debug.o mutt/exit.o \
+LIBMUTTOBJS=	mutt/address.o mutt/base64.o mutt/buffer.o mutt/charset.o \
+		mutt/date.o mutt/debug.o mutt/exit.o \
 		mutt/file.o mutt/hash.o mutt/idna.o mutt/list.o \
 		mutt/mapping.o mutt/mbyte.o mutt/md5.o \
 		mutt/memory.o mutt/message.o mutt/mime.o mutt/parameter.o \
@@ -267,14 +267,14 @@ git_ver.h: $(ALL_FILES)
 	cmp -s git_ver.h.tmp git_ver.h || mv git_ver.h.tmp git_ver.h; \
 	rm -f git_ver.h.tmp
 
-hcache/hcversion.h:	$(SRCDIR)/mutt.h $(SRCDIR)/address.h \
+hcache/hcversion.h:	$(SRCDIR)/mutt.h $(SRCDIR)/mutt/address.h \
 			$(SRCDIR)/mutt/list.h $(SRCDIR)/mutt/buffer.h \
 			$(SRCDIR)/mutt/parameter.h $(SRCDIR)/body.h \
 			$(SRCDIR)/envelope.h $(SRCDIR)/header.h \
 			$(SRCDIR)/hcache/hcachever.sh \
 			$(PWD)/hcache
 	( echo '#include "config.h"'; echo '#include "mutt.h"'; \
-	echo '#include "address.h"'; echo '#include "mutt/list.h"'; \
+	echo '#include "mutt/address.h"'; echo '#include "mutt/list.h"'; \
 	echo '#include "mutt/buffer.h"'; echo '#include "mutt/parameter.h"'; \
 	echo '#include "body.h"'; echo '#include "envelope.h"'; \
 	echo '#include "header.h"';) | $(CPP) $(CFLAGS) - | \
