+++ alias.c
@@ -32,7 +32,6 @@
 #include "mutt/mutt.h"
 #include "mutt.h"
 #include "alias.h"
-#include "address.h"
 #include "envelope.h"
 #include "globals.h"
 #include "mutt_curses.h"
