+++ copy.c
@@ -28,7 +28,6 @@
 #include "mutt/mutt.h"
 #include "mutt.h"
 #include "copy.h"
-#include "address.h"
 #include "body.h"
 #include "context.h"
 #include "envelope.h"
