+++ hdrline.c
@@ -31,7 +31,6 @@
 #include <time.h>
 #include "mutt/mutt.h"
 #include "mutt.h"
-#include "address.h"
 #include "body.h"
 #include "context.h"
 #include "envelope.h"
