+++ mbox.c
@@ -34,7 +34,6 @@
 #include <utime.h>
 #include "mutt/mutt.h"
 #include "mutt.h"
-#include "address.h"
 #include "body.h"
 #include "buffy.h"
 #include "context.h"
