+++ mutt/address.c
@@ -57,13 +57,17 @@
  * | mutt_addr_valid_msgid()      | Is this a valid Message ID?
  * | mutt_addr_write()            | Write an Address to a buffer
  * | mutt_addr_write_single()     | Write a single Address to a buffer
+ * | mutt_addrlist_to_intl()      | Convert an Address list to Punycode
+ * | mutt_addrlist_to_local()     | Convert an Address list from Punycode
  */
 
 #include "config.h"
 #include <stdio.h>
 #include <string.h>
-#include "mutt/mutt.h"
 #include "address.h"
+#include "idna2.h"
+#include "memory.h"
+#include "string2.h"
 
 /**
  * AddressSpecials - Characters with special meaning for email addresses
@@ -1230,4 +1234,69 @@ size_t mutt_addr_write(char *buf, size_t buflen, struct Address *addr, bool disp
 done:
   *pbuf = 0;
   return pbuf - buf;
+}
+
+/**
+ * mutt_addrlist_to_intl - Convert an Address list to Punycode
+ * @param[in]  a   Address list to modify
+ * @param[out] err Pointer for failed addresses
+ * @retval 0  Success, all addresses converted
+ * @retval -1 Error, err will be set to the failed address
+ */
+int mutt_addrlist_to_intl(struct Address *a, char **err)
+{
+  char *user = NULL, *domain = NULL;
+  char *intl_mailbox = NULL;
+  int rc = 0;
+
+  if (err)
+    *err = NULL;
+
+  for (; a; a = a->next)
+  {
+    if (!a->mailbox || mutt_addr_is_intl(a))
+      continue;
+
+    if (mutt_addr_mbox_to_udomain(a->mailbox, &user, &domain) == -1)
+      continue;
+
+    intl_mailbox = mutt_idna_local_to_intl(user, domain);
+    if (!intl_mailbox)
+    {
+      rc = -1;
+      if (err && !*err)
+        *err = mutt_str_strdup(a->mailbox);
+      continue;
+    }
+
+    mutt_addr_set_intl(a, intl_mailbox);
+  }
+
+  return rc;
+}
+
+/**
+ * mutt_addrlist_to_local - Convert an Address list from Punycode
+ * @param a Address list to modify
+ * @retval 0 Always
+ */
+int mutt_addrlist_to_local(struct Address *a)
+{
+  char *user = NULL, *domain = NULL;
+  char *local_mailbox = NULL;
+
+  for (; a; a = a->next)
+  {
+    if (!a->mailbox || mutt_addr_is_local(a))
+      continue;
+
+    if (mutt_addr_mbox_to_udomain(a->mailbox, &user, &domain) == -1)
+      continue;
+
+    local_mailbox = mutt_idna_intl_to_local(user, domain, 0);
+    if (local_mailbox)
+      mutt_addr_set_local(a, local_mailbox);
+  }
+
+  return 0;
 }
