+++ muttlib.c
@@ -42,7 +42,6 @@
 #include <wchar.h>
 #include "mutt/mutt.h"
 #include "mutt.h"
-#include "address.h"
 #include "alias.h"
 #include "body.h"
 #include "envelope.h"
