+++ mx.c
@@ -33,7 +33,6 @@
 #include "mutt/mutt.h"
 #include "mutt.h"
 #include "mx.h"
-#include "address.h"
 #include "body.h"
 #include "buffy.h"
 #include "context.h"
