+++ ncrypt/crypt_gpgme.c
@@ -44,7 +44,6 @@
 #include <unistd.h>
 #include "mutt/mutt.h"
 #include "mutt.h"
-#include "address.h"
 #include "alias.h"
 #include "body.h"
 #include "crypt.h"
