+++ ncrypt/pgp.c
@@ -43,7 +43,6 @@
 #include "mutt/mutt.h"
 #include "mutt.h"
 #include "pgp.h"
-#include "address.h"
 #include "body.h"
 #include "crypt.h"
 #include "cryptglue.h"
