+++ ncrypt/pgpinvoke.c
@@ -31,7 +31,6 @@
 #include <string.h>
 #include <unistd.h>
 #include "mutt/mutt.h"
-#include "address.h"
 #include "filter.h"
 #include "format_flags.h"
 #include "globals.h"
