+++ ncrypt/pgpkey.c
@@ -35,7 +35,6 @@
 #include "mutt/mutt.h"
 #include "mutt.h"
 #include "pgpkey.h"
-#include "address.h"
 #include "body.h"
 #include "crypt.h"
 #include "filter.h"
