+++ recvcmd.c
@@ -27,7 +27,6 @@
 #include <string.h>
 #include "mutt/mutt.h"
 #include "mutt.h"
-#include "address.h"
 #include "alias.h"
 #include "attach.h"
 #include "body.h"
