+++ remailer.c
@@ -34,7 +34,6 @@
 #include "mutt/mutt.h"
 #include "mutt.h"
 #include "remailer.h"
-#include "address.h"
 #include "envelope.h"
 #include "filter.h"
 #include "format_flags.h"
